New Recipes to Fix Sapling Conversions
I just made a new recipe with the same ratios to allow oak to be converted to each type of sapling, to make the behavior consistent with what people experienced in 1.12.

If the change was intentional, feel free to let me know. But it seemed to just be another issue with block variants being converted from damage values to actual independent blocks.

If there is a way to get the same functionality we had in 1.12.2, where you could just put any sapling in, and it would auto convert it to oak, without having to select a specific recipe, that'd be ideal, but as far as I know, because there are no more damage value varients like there used to be, that isn't currently possible (Within the limits of my admittedly shallow understanding of factorymod)

Charcoal also likely has this same issue, because currently, it you can only convert oak into charcoal from what I'm hearing from people in the server. If I fix that today, I will make it as a new pull request however.