Remove mountains overlapping with cold
Overlapping biomes in multiple categories cause conflicts in growth times, it is better to have them separate and just list the separate categories for each crop. In this case every crop with cold already had mountain listed, except for pumpkins so I just added it in. Overall goal of this commit is to fix potatoes which had different growth rates for cold and mountains, and since mountains were listed in both categories they were getting the wrong growth rate.