Add in multiple layers of enderchests relay for shops
I don't get why this is a setting, already coded, and not used?

It allows for actual wealth investment in shops, even if you relay a shop chest 3times, you still need to invest 12d in the glass, 3d in the enderchests, and then surround that relay with even more layers of DRO to make it worthwhile. (60d in obby on first layer, and you'd want each successive relay to have another additional layer of dro on it) A single secure relay system could *easily* cost 300d.